requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
python-dotenv>=1.0.1

requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.1.0
tqdm>=4.66.0

pandas>=2.0.0
openpyxl>=3.1.0

numpy>=1.24.0
scikit-learn>=1.3.0
rank-bm25>=0.2.2
sentence-transformers>=2.6.0

google-generativeai>=0.5.0

streamlit>=1.31.0
